# ai-test-agent/examples/openapi_demo.yaml
openapi: 3.0.0
info:
  title: Demo 电商服务
  version: 1.0.0
paths:
  /api/v1/login:
    post:
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: {type: string}
                password: {type: string}
      responses:
        '200':
          description: 登录成功，返回 token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: {type: string}
                  user: {type: string}
        '401':
          description: 用户名或密码错误
  /api/v1/products:
    post:
      summary: 创建商品
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, price]
              properties:
                name: {type: string, maxLength: 100}
                price: {type: number, minimum: 0, exclusiveMinimum: true}
      responses:
        '201':
          description: 创建成功
        '422':
          description: 参数校验失败
  /api/v1/products/{product_id}:
    get:
      summary: 查询商品
      parameters:
        - name: product_id
          in: path
          required: true
          schema: {type: integer}
      responses:
        '200': {description: 商品信息}
        '404': {description: 商品不存在}
    put:
      summary: 更新商品
      parameters:
        - name: product_id
          in: path
          required: true
          schema: {type: integer}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, price]
              properties:
                name: {type: string, maxLength: 100}
                price: {type: number, minimum: 0, exclusiveMinimum: true}
      responses:
        '200': {description: 更新成功}
        '404': {description: 商品不存在}
        '422': {description: 参数校验失败}
    delete:
      summary: 删除商品
      parameters:
        - name: product_id
          in: path
          required: true
          schema: {type: integer}
      responses:
        '204': {description: 删除成功}
        '404': {description: 商品不存在}